Make the local data boolean a instance variable
Looks like the intent given to the chooser is not returned in onActivitiyResult, so the
getExtras called in onActivityResult is always null and the check for includeData always
fails.

So we make the includeData a instance variable of the plugin, set it on execute and
read it in the onActivityResult method.